Clean up JS for minor idioms, and Crockford style.
Also: added comments to a few places where things might be unobvious
to them--particularly the use of "var self = this", to allow a closure
in the setTimeout() function.

Changed the signature of showTempResultMessage() to take the
msg-to-be-set into this and actually manipulate the DOM here, rather
than in init. This seemed like this would be a useful refactoring for
the additional work they have to do afterwards.